+++ CHANGELOG.md
@@ -2,6 +2,11 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
+_(none)_
+
+---
+
+## 3.0.0-rc1 (2023-07-12)
 * Breaking: change `CONFIG:spread` to `CONFIG:enable-define`
 * Breaking: tokens by value (prefixed with `$`) always resolve to a value
 * Breaking: CSS variable references (`var`) always fallback to a value
@@ -12,8 +17,6 @@ CHANGELOG
 * Feature: lists can now be defined with references
 * Feature: `apply` complex tokens
 * Feature: `basename` function
-
----
 
 ## 2.0.0 (2023-06-11)